+++ crates/services/executor/src/executor.rs
@@ -116,6 +116,7 @@ use fuel_core_types::{
         block_producer::Components,
         executor::{
             Error as ExecutorError,
+            Event as ExecutorEvent,
             ExecutionKind,
             ExecutionResult,
             ExecutionType,
@@ -262,6 +263,7 @@ pub struct ExecutionData {
     found_mint: bool,
     message_ids: Vec<MessageId>,
     tx_status: Vec<TransactionExecutionStatus>,
+    events: Vec<ExecutorEvent>,
     pub skipped_transactions: Vec<(TxId, ExecutorError)>,
 }
 
@@ -483,6 +485,7 @@ where
             message_ids,
             tx_status,
             skipped_transactions,
+            events,
             ..
         } = execution_data;
 
@@ -511,6 +514,7 @@ where
             block,
             skipped_transactions,
             tx_status,
+            events,
         };
 
         // Get the complete fuel block.
@@ -534,6 +538,7 @@ where
             found_mint: false,
             message_ids: Vec::new(),
             tx_status: Vec::new(),
+            events: Vec::new(),
             skipped_transactions: Vec::new(),
         };
         let execution_data = &mut data;
@@ -546,7 +551,7 @@ where
         let block_height = *block.header.height();
 
         if self.relayer.enabled() {
-            self.process_da(block_st_transaction, &block.header)?;
+            self.process_da(block_st_transaction, &block.header, execution_data)?;
         }
 
         // ALl transactions should be in the `TxSource`.
@@ -659,6 +664,7 @@ where
         &self,
         block_st_transaction: &mut D,
         header: &PartialBlockHeader,
+        execution_data: &mut ExecutionData,
     ) -> ExecutorResult<()> {
         let block_height = *header.height();
         let prev_block_height = block_height
@@ -689,6 +695,9 @@ where
                         block_st_transaction
                             .storage::<Messages>()
                             .insert(&message.nonce, &message)?;
+                        execution_data
+                            .events
+                            .push(ExecutorEvent::NewMessage(message));
                     }
                 }
             }
@@ -853,7 +862,7 @@ where
 
             self.persist_output_utxos(
                 block_height,
-                execution_data.tx_count,
+                execution_data,
                 &coinbase_id,
                 block_st_transaction.as_mut(),
                 inputs.as_slice(),
@@ -1025,12 +1034,17 @@ where
         }
 
         // change the spent status of the tx inputs
-        self.spend_input_utxos(tx.inputs(), tx_st_transaction.as_mut(), reverted)?;
+        self.spend_input_utxos(
+            tx.inputs(),
+            tx_st_transaction.as_mut(),
+            reverted,
+            execution_data,
+        )?;
 
         // Persist utxos first and after calculate the not utxo outputs
         self.persist_output_utxos(
             *header.height(),
-            execution_data.tx_count,
+            execution_data,
             &tx_id,
             tx_st_transaction.as_mut(),
             tx.inputs(),
@@ -1201,25 +1215,52 @@ where
         inputs: &[Input],
         db: &mut D,
         reverted: bool,
+        execution_data: &mut ExecutionData,
     ) -> ExecutorResult<()> {
         for input in inputs {
             match input {
-                Input::CoinSigned(CoinSigned { utxo_id, .. })
-                | Input::CoinPredicate(CoinPredicate { utxo_id, .. }) => {
+                Input::CoinSigned(CoinSigned {
+                    utxo_id,
+                    owner,
+                    amount,
+                    asset_id,
+                    maturity,
+                    ..
+                })
+                | Input::CoinPredicate(CoinPredicate {
+                    utxo_id,
+                    owner,
+                    amount,
+                    asset_id,
+                    maturity,
+                    ..
+                }) => {
                     // prune utxo from db
-                    db.storage::<Coins>().remove(utxo_id)?;
+                    let coin = db
+                        .storage::<Coins>()
+                        .remove(utxo_id)?
+                        .map(Ok)
+                        .unwrap_or_else(|| {
+                            // If the coin is not found in the database, it means that it was
+                            // already spent or `utxo_validation` is `false`.
+                            self.get_coin_or_default(
+                                db, *utxo_id, *owner, *amount, *asset_id, *maturity,
+                            )
+                        })?;
+
+                    execution_data.events.push(ExecutorEvent::ConsumeCoin(coin));
                 }
-                Input::MessageDataSigned(_)
-                | Input::MessageDataPredicate(_)
-                if reverted => {
+                Input::MessageDataSigned(_) | Input::MessageDataPredicate(_)
+                    if reverted =>
+                {
                     // Don't spend the retryable messages if transaction is reverted
                     continue
                 }
                 Input::MessageCoinSigned(MessageCoinSigned { nonce, .. })
                 | Input::MessageCoinPredicate(MessageCoinPredicate { nonce, .. })
-                | Input::MessageDataSigned(MessageDataSigned { nonce, .. }) // Spend only if tx is not reverted
-                | Input::MessageDataPredicate(MessageDataPredicate { nonce, .. }) // Spend only if tx is not reverted
-                => {
+                | Input::MessageDataSigned(MessageDataSigned { nonce, .. })
+                | Input::MessageDataPredicate(MessageDataPredicate { nonce, .. }) => {
+                    // `MessageDataSigned` and `MessageDataPredicate` are spend only if tx is not reverted
                     // mark message id as spent
                     let was_already_spent =
                         db.storage::<SpentMessages>().insert(nonce, &())?;
@@ -1228,7 +1269,13 @@ where
                         return Err(ExecutorError::MessageAlreadySpent(*nonce))
                     }
                     // cleanup message contents
-                    db.storage::<Messages>().remove(nonce)?;
+                    let message = db
+                        .storage::<Messages>()
+                        .remove(nonce)?
+                        .ok_or_else(|| ExecutorError::MessageAlreadySpent(*nonce))?;
+                    execution_data
+                        .events
+                        .push(ExecutorEvent::ConsumeMessage(message));
                 }
                 _ => {}
             }
@@ -1523,12 +1570,13 @@ where
     fn persist_output_utxos(
         &self,
         block_height: BlockHeight,
-        tx_idx: u16,
+        execution_data: &mut ExecutionData,
         tx_id: &Bytes32,
         db: &mut D,
         inputs: &[Input],
         outputs: &[Output],
     ) -> ExecutorResult<()> {
+        let tx_idx = execution_data.tx_count;
         for (output_index, output) in outputs.iter().enumerate() {
             let index = u8::try_from(output_index)
                 .expect("Transaction can have only up to `u8::MAX` outputs");
@@ -1540,7 +1588,7 @@ where
                     to,
                 } => Self::insert_coin(
                     block_height,
-                    tx_idx,
+                    execution_data,
                     utxo_id,
                     amount,
                     asset_id,
@@ -1570,7 +1618,7 @@ where
                     amount,
                 } => Self::insert_coin(
                     block_height,
-                    tx_idx,
+                    execution_data,
                     utxo_id,
                     amount,
                     asset_id,
@@ -1583,7 +1631,7 @@ where
                     amount,
                 } => Self::insert_coin(
                     block_height,
-                    tx_idx,
+                    execution_data,
                     utxo_id,
                     amount,
                     asset_id,
@@ -1606,7 +1654,7 @@ where
 
     fn insert_coin(
         block_height: BlockHeight,
-        tx_idx: u16,
+        execution_data: &mut ExecutionData,
         utxo_id: UtxoId,
         amount: &Word,
         asset_id: &AssetId,
@@ -1622,13 +1670,14 @@ where
                 amount: *amount,
                 asset_id: *asset_id,
                 maturity: 0u32.into(),
-                tx_pointer: TxPointer::new(block_height, tx_idx),
+                tx_pointer: TxPointer::new(block_height, execution_data.tx_count),
             }
             .into();
 
             if db.storage::<Coins>().insert(&utxo_id, &coin)?.is_some() {
                 return Err(ExecutorError::OutputAlreadyExists)
             }
+            execution_data.events.push(ExecutorEvent::NewCoin(coin));
         }
 
         Ok(())